fix: slip up schema into separate files
