+++ packages/db/schema/client.ts
@@ -1,0 +1,21 @@
+import { sql } from "drizzle-orm";
+import { serial, timestamp, varchar } from "drizzle-orm/mysql-core";
+
+import { mySqlTable } from "./_table";
+
+export const clients = mySqlTable("clients", {
+  id: serial("id").primaryKey(),
+  name: varchar("name", { length: 256 }).notNull(),
+  email: varchar("email", { length: 256 }),
+  phoneNumber: varchar("phoneNumber", { length: 15 }),
+  createdAt: timestamp("created_at")
+    .default(sql`CURRENT_TIMESTAMP`)
+    .notNull(),
+  updatedAt: timestamp("updated_at").onUpdateNow(),
+  profile: varchar("profile", { length: 256 }),
+  address: varchar("address", { length: 256 }),
+  zip: varchar("zip", { length: 256 }),
+  city: varchar("city", { length: 256 }),
+  state: varchar("state", { length: 256 }),
+  authId: varchar("auth_id", { length: 256 }),
+});
